Revert "minor update - fix customer name and move ContactCustomer class into …"
