+++ config/amoCRM.php
@@ -3,6 +3,6 @@
 return [
     'client_id'           => 'e095090d-d4d5-4202-aabc-cd0f3f40532a',
     'client_secret'       => 'DUx6xIhMA6FhIRJlLPXzJkKvLuwC6Chxb9R3ffYAzKfRpfgY41pL0NZAZRq1NOtV',
-    'client_redirect_url' => 'https://5d04-94-158-52-23.ngrok-free.app/api/get-credentials',
+    'client_redirect_url' => 'https://ca80-213-230-100-58.ngrok-free.app/api/get-credentials',
     'base_domain'         => 'anurillakhonov.amocrm.ru'
 ];
